remove globals
Lox uses globals to enable mutual recursive functions and in general
functions declared out of order. This has advantages in REPL
environment but makes every undefined variable usage a runtime error
even for scripts where every name should be resolvable at compile time.

We want to focus more on static scripts and want to eliminate these
trivial runtime errors. A replacement for the global variable will be a
different kind of binding that will be stored on the stack with other
locals but will be resolved in a separate pass before normal `let`
bindings. This binding type will be used for function items and maybe
some kind of `let rec` or `const` binding.

These "rec" bindings will be resolvable lexically before they're defined
but there will have to be some restrictions placed on them because of
that. One is that they can't be shadowed within the same scope, and
there will probably have to be some compile-time cycle detection for
data dependency, unless we want to get into lazy evaluation later.

```
let rec a = b;
let rec b = a;
```

This may still change a lot, but we definitely don't want globals.